apiPython/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]==3.1.12
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6